LayoutTests/fast/dom/Selection/wrong-document-err-expected.txt

Test that collapse() and selectAllChildren() throw WRONG_DOCUMENT_ERR if the node is in another document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sel.collapse(externalTextNode, 0) threw exception Error: WRONG_DOCUMENT_ERR: DOM Exception 4.
PASS sel.selectAllChildren(externalTextNode) threw exception Error: WRONG_DOCUMENT_ERR: DOM Exception 4.
PASS sel.collapse(internalTextNode, 0) is undefined
PASS sel.selectAllChildren(internalTextNode) is undefined
PASS successfullyParsed is true

TEST COMPLETE
